7. ignore controller test 를 기존 user controller 테스트에서 분리
